restructured auth and users
