+++ package.json
@@ -14,12 +14,15 @@
     "cross-env": "7.0.3",
     "dotenv": "^16.0.3",
     "express": "^4.17.1",
+    "gravatar": "^1.8.2",
     "http-errors": "^2.0.0",
+    "jimp": "^0.16.2",
     "joi": "^17.7.0",
     "jsonwebtoken": "^9.0.0",
     "mongodb": "^4.13.0",
     "mongoose": "^6.8.3",
     "morgan": "^1.10.0",
+    "multer": "^1.4.5-lts.1",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "uuid": "^9.0.0"
@@ -31,6 +34,7 @@
     "eslint-plugin-import": "2.25.3",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "5.2.0",
+    "jest": "^29.4.0",
     "nodemon": "2.0.15"
   }
 }
